Reame: 경로 설정 및 폴더 생성
