+++ moamoa/src/Components/Post/PostCardItem.jsx
@@ -1,28 +1,24 @@
 import React, { useState } from 'react';
 import { Link } from 'react-router-dom';
 import PostCardUser from './PostCardUser';
-import MoreBtn from './MoreBtn';
+import MoreBtn from '../Common/MoreBtn';
 import styled from 'styled-components';
 import heartBg from '../../Assets/icons/heart.svg';
 import heartBgFill from '../../Assets/icons/heart-fill.svg';
 import commentBg from '../../Assets/icons/message-circle.svg';
+import Datacalc from '../Common/datecalc'
 
-export default function PostCardHome(post) {
+export default function PostCardDetail(post) {
+  // const [heartCount, setHeartCount] = useState("");
   const [toggleCount, setToggleCount] = useState(false);
   const [heartcolor, setHeartColor] = useState(heartBg);
 
-  const postprop = post.post;
-  const profileImgUrl = `${postprop.author.image}`;
+  // const baseUrl = `https://api.mandarin.weniv.co.kr/`
+
+  const postprop = post.post.post;
   const postImgUrl = `${postprop.image}`;
   const postDetailId = post.post.id;
   const postDetailUrl = `/post/${postDetailId}`;
-
-  const inputDate = postprop.createdAt;
-  const dateset = inputDate.split('').slice(0, 10).join('');
-  const year = dateset.slice(0, 4);
-  const month = dateset.slice(5, 7);
-  const day = dateset.slice(8, 10);
-  const outputDate = `${year}년 ${month}월 ${day}일`;
 
   const handleHeartCount = () => {
     if (toggleCount === true) {
@@ -39,7 +35,7 @@ export default function PostCardHome(post) {
           <PostArticle>
             <Frofile>
               <PostCardUser
-                url={profileImgUrl}
+                url={postprop.author.image}
                 username={postprop.author.username}
                 accountname={postprop.author.accountname}
               />
@@ -50,7 +46,7 @@ export default function PostCardHome(post) {
               {postImgUrl ? <PostImg src={postImgUrl} alt='게시글 사진' /> : null}
             </Link>
             <PostFooterContainer>
-              <CreateDate>{outputDate}</CreateDate>
+              <CreateDate>{Datacalc(postprop.createdAt)}</CreateDate>
               <div>
                 <HeartBtn
                   onClick={() => {
